+++ resources/views/auth/login-two-factor.blade.php
@@ -3,13 +3,17 @@
         {{ __('Authenticate with your code') }}
     </h2>
     @if ($twoFactorType === 'email' || $twoFactorType === 'phone')
-        <div wire:poll.5s>
-            {{ $this->resend }}
-        </div>
+    <div wire:poll.5s>
+        {{ $this->resend }}
+    </div>
     @endif
     <form method="POST" action="{{ route('two-factor.login') }}" class="space-y-8">
-
         @csrf
+        
+        <div style="display: none">
+            <input type="text" id="recovery_code" name="recovery_code" value="">
+        </div>
+
         {{ $this->form }}
 
         <div class="flex items-center justify-between mt-6">