+++ src/Http/Livewire/Auth/PasswordReset.php
@@ -77,7 +77,7 @@ class PasswordReset extends Page implements HasForms
 
     public function render(): View
     {
-        return view('filament-two-factor-auth::auth.password-reset', $this->getViewData())
+        return view('filament-2fa::auth.password-reset', $this->getViewData())
             ->layout($this->getLayout(), [
                 'livewire' => $this,
                 'maxContentWidth' => $this->getMaxContentWidth(),