+++ src/Notifications/SendOTP.php
@@ -30,9 +30,9 @@ class SendOTP extends Notification implements ShouldQueue
 
         if (
             $notifiable->two_factor_type === TwoFactorType::phone &&
-            in_array(TwoFactorType::phone, config('filament-two-factor-auth.options'))
+            in_array(TwoFactorType::phone, config('filament-2fa.options'))
         ) {
-            return [config('filament-two-factor-auth.sms_service')];
+            return [config('filament-2fa.sms_service')];
         }
 
         return [];